backend/requirements.txt
# Core ML and Transformers
transformers>=4.36.0
torch>=2.1.0
torchvision>=0.15.0
torchaudio>=2.0.0
accelerate>=0.25.0
bitsandbytes>=0.41.1
safetensors>=0.4.0
diffusers>=0.24.0
tokenizers>=0.15.0
protobuf>=4.25.1

# System and Monitoring
psutil>=5.9.0
wmi>=1.5.1
nvidia-ml-py3>=7.352.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
loguru>=0.7.2
typing-extensions>=4.10.0
aiohttp>=3.9.5
beautifulsoup4>=4.9.3
nest-asyncio>=1.5.8

# Data Processing and Visualization
numpy>=1.24.3
pandas>=2.1.3
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
scikit-learn>=1.3.2
sentence-transformers>=2.2.2

# News API and Search
newsapi-python>=0.2.7
duckduckgo-search>=4.1.1

# Vector Database
faiss-cpu>=1.7.4

# Image Processing
pillow>=10.1.0

# Translation
deep-translator>=1.11.4

# FastAPI and related packages
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Add these lines to your requirements.txt
ftfy>=6.1.1
regex>=2023.8.8
clip @ git+https://github.com/openai/CLIP.git

